Refactor: Fully isolate horse racing game
This commit addresses feedback to fully isolate the horse racing game. The `race_day_scheduler.py` has been moved back to the root directory, and the `game_manager` is no longer imported or set up in `bot.py`, making the `horse_racing_game` directory completely self-contained.
