+++ horse_racing_game/configs/bot_personalities.json
