+++ horse_racing_game/configs/horse_attributes.json
